+++ src/spyglass/data_import/__init__.py
@@ -1,1 +1,2 @@
+# TODO: change naming to avoid match between module and function
 from .insert_sessions import insert_sessions