+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add `deprecation_factory` to facilitate table migration. #717
 - Add Spyglass logger. #730
 - IntervalList: Add secondary key `pipeline` #742
+- Increase pytest coverage for `common`, `lfp`, and `utils`. #743
 
 ### Pipelines
 
@@ -25,12 +26,11 @@
     - Refactor input validation in DLC pipeline. #688
     - DLC path handling from config, and normalize naming convention. #722
 - Decoding:
-    - Add `decoding` pipeline V1. #731
+    - Add `decoding` pipeline V1. #731, #769
     - Add a table to store the decoding results #731
     - Use the new `non_local_detector` package for decoding #731
     - Allow multiple spike waveform features for clusterelss decoding #731
     - Reorder notebooks #731
-
 
 ## [0.4.3] (November 7, 2023)